+++ core/src/org/eamonn/trog/inGameUserInterface.scala
@@ -179,14 +179,14 @@ object inGameUserInterface {
     var path = Pathfinding.findPath(game.player.location, loc, game.level)
     if (path.isEmpty || !game.explored.contains(loc)) {
       batch.setColor(1f, 0f, 0f, .75f)
-      Animation.twoFrameAnimation(game, batch, "mouseHover", loc.x, loc.y)
+      Animation.twoFrameAnimation(game, batch, "mouseHover", loc.x.toFloat, loc.y.toFloat)
     } else {
       path.foreach(p => {
         batch.setColor(1f, 1f, 1f, .75f)
-        Animation.twoFrameAnimation(game, batch, "mouseHover", loc.x, loc.y)
+        Animation.twoFrameAnimation(game, batch, "mouseHover", loc.x.toFloat, loc.y.toFloat)
         p.list.reverse.tail.foreach(l => {
           batch.setColor(1f, 1f, 1f, .75f)
-          Animation.twoFrameAnimation(game, batch, "pathTrail", l.x, l.y)
+          Animation.twoFrameAnimation(game, batch, "pathTrail", l.x.toFloat, l.y.toFloat)
         })
       })
     }
